+++ src/api/queryClient.ts
@@ -1,18 +1,11 @@
 import {
   QueryClient,
   QueryClientConfig,
-  QueryFunction,
-  QueryKey,
   useQuery,
   UseQueryOptions
 } from "@tanstack/react-query";
 import { useEnvironment } from "../environment/environmentContext";
-export const fetchData = (path: string) => fetch(`${process.env.API_URL}${path}`);
 
-export const fetchJson = <T extends unknown>(path: string) => fetchData(path).then(res => res.json() as T)
-
-export const fetcher = <T>({ queryKey }: { queryKey: readonly unknown[] }): Promise<T> =>
-  fetchData(`${queryKey[0]}`).then((res) => res.json());
 
 export const queryClientOptions: QueryClientConfig = {
   defaultOptions: {
