+++ src/app/[projectID]/layout.tsx
@@ -5,6 +5,9 @@ import React from "react";
 import { fetchProjectConfig } from "../../api/projectConfig/projectConfig";
 import environment from "../../environment/server";
 import { ProjectPageT } from "../../types";
+import { defaultMetadata } from "../../utils/seo";
+
+export const metadata = defaultMetadata
 
 const Layout: ProjectPageT = async ({children, params}) => {
   const projectConfig = await fetchProjectConfig(params.projectID)
