+++ src/app/[projectID]/map/layout.tsx
@@ -2,6 +2,12 @@ import URLModifier from "./_components/URLModifier";
 import QueryField from "./_components/QueryField";
 import styles from './_styles/layout.module.css'
 import { SelectedSchoolProvider } from "../../../hooks/useSelectedSchool";
+import createProjectMetadata from "../../../utils/seo";
+
+export const generateMetadata = createProjectMetadata((projectName) => ({
+  title: 'Mapa',
+  description: `Wyszukaj szkoły średnie z po8klasie ${projectName}`
+}))
 
 const MapSearchPageLayout = ({children, map, listing}) => (
   <SelectedSchoolProvider>
